pyproject.toml
[project]
name = "antigravity-os"
version = "2026.2"
description = "抵抗认知熵增 — 自动将信息流转化为结构化知识资产"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "httpx>=0.24",
    "requests>=2.28",
    "feedparser>=6.0",
    "beautifulsoup4>=4.12",
    "pyyaml>=6.0",
    "rich>=13.0",
    "openai>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["agos*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
